requirements.txt
# QBO Copilot Dependencies

# HTTP & API
requests>=2.31.0

# OAuth / Environment
python-dotenv>=1.0.0

# Configuration
pyyaml>=6.0

# Slack Integration
slack-bolt>=1.18.0
slack-sdk>=3.23.0

# Google Drive Integration
google-api-python-client>=2.0
google-auth>=2.0

# AI / LLM
anthropic>=0.18.0

# WhatsApp (future)
# httpx>=0.25.0

# OpenClaw / Agent Framework
# openclaw>=0.1.0  # Install when available

# Development
pytest>=7.4.0
black>=23.0.0
